fixed payment methods soft-delete related issues, removed subscription payment methods soft-delete filtering, added customer addresses endpoint query, added customer by id endpoint - DRUMEO-953, DEVE-104, DEVE-111, DEVE-109

+++ src/Controllers/PaymentMethodJsonController.php
@@ -632,6 +632,13 @@ class PaymentMethodJsonController extends Controller
             );
         }
 
+        throw_if(
+            !empty($paymentMethod->getDeletedAt()),
+            new NotAllowedException(
+                'Payment method already soft-deleted'
+            )
+        );
+
         $this->entityManager->remove($paymentMethod);
         $this->entityManager->flush();
